TWINFACES-258 [input field] class (single)
