--- src/features/twinClass/selectField/index.ts
@@ -1,1 +1,0 @@
-export * from "./twin-class-select-field";
