+++ src/screens/twin-class-relations/form-fields.tsx
@@ -1,10 +1,10 @@
-import { ComboboxFormField } from "@/components/form-fields/combobox";
 import { TextFormField } from "@/components/form-fields";
+import { ComboboxFormField } from "@/components/form-fields/combobox";
 import {
   useTwinClassLinkStrengthSelectAdapter,
   useTwinClassLinkTypeSelectAdapter,
 } from "@/entities/twin-class-link";
-import { TwinClassSelectField } from "@/features/twinClass";
+import { useTwinClassSelectAdapter } from "@/entities/twinClass";
 import { Control, FieldValues, Path } from "react-hook-form";
 
 export function TwinClassRelationsFormFields<T extends FieldValues>({
@@ -14,22 +14,31 @@ export function TwinClassRelationsFormFields<T extends FieldValues>({
   control: Control<T>;
   isForward?: boolean;
 }) {
+  const tcAdapter = useTwinClassSelectAdapter();
   const typeAdapter = useTwinClassLinkTypeSelectAdapter();
   const strengthAdapter = useTwinClassLinkStrengthSelectAdapter();
 
   return (
     <>
-      <TwinClassSelectField
+      <ComboboxFormField
         control={control}
         name={"srcTwinClassId" as Path<T>}
         label="Source Twin Class"
         disabled={isForward}
+        selectPlaceholder="Select twin class"
+        searchPlaceholder="Search twin class..."
+        noItemsText={"No classes found"}
+        {...tcAdapter}
       />
-      <TwinClassSelectField
+      <ComboboxFormField
         control={control}
         name={"dstTwinClassId" as Path<T>}
         label="Destination Twin Class"
         disabled={!isForward}
+        selectPlaceholder="Select twin class"
+        searchPlaceholder="Search twin class..."
+        noItemsText={"No classes found"}
+        {...tcAdapter}
       />
       <TextFormField
         control={control}
